🔨 refacto: move intents lib into sub folder
