+++ src/index.js
@@ -8,7 +8,7 @@ import * as data from './data'
 import * as cozyFetch from './fetch'
 import * as mango from './mango'
 import * as files from './files'
-import * as intents from './intents'
+import * as intents from './intents/'
 import * as jobs from './jobs'
 import * as offline from './offline'
 import * as settings from './settings'
